+++ sha256_mb/sha256_mb_test.c
@@ -188,7 +188,7 @@ main(void)
                                             HASH_ENTIRE);
 
                 if (ctx) {
-                        t = (unsigned long) (ctx->user_data);
+                        t = (uintptr_t) (ctx->user_data);
                         good = expResultDigest[t];
                         checked++;
                         for (j = 0; j < SHA256_DIGEST_NWORDS; j++) {
@@ -212,7 +212,7 @@ main(void)
                 ctx = sha256_ctx_mgr_flush(mgr);
 
                 if (ctx) {
-                        t = (unsigned long) (ctx->user_data);
+                        t = (uintptr_t) (ctx->user_data);
                         good = expResultDigest[t];
                         checked++;
                         for (j = 0; j < SHA256_DIGEST_NWORDS; j++) {
@@ -248,7 +248,7 @@ main(void)
                 ctx = sha256_ctx_mgr_submit(mgr, &ctxpool[i], msgs[j], strlen((char *) msgs[j]),
                                             HASH_ENTIRE);
                 if (ctx) {
-                        t = (unsigned long) (ctx->user_data);
+                        t = (uintptr_t) (ctx->user_data);
                         k = PSEUDO_RANDOM_NUM(t);
                         good = expResultDigest[k];
                         checked++;
@@ -267,7 +267,7 @@ main(void)
                                 goto end;
                         }
 
-                        t = (unsigned long) (ctx->user_data);
+                        t = (uintptr_t) (ctx->user_data);
                         k = PSEUDO_RANDOM_NUM(t);
                 }
         }
@@ -275,7 +275,7 @@ main(void)
                 ctx = sha256_ctx_mgr_flush(mgr);
 
                 if (ctx) {
-                        t = (unsigned long) (ctx->user_data);
+                        t = (uintptr_t) (ctx->user_data);
                         k = PSEUDO_RANDOM_NUM(t);
                         good = expResultDigest[k];
                         checked++;
